+++ CHANGELOG.md
@@ -7,10 +7,11 @@ Full documentation for rocLibrary is available at [https://rocm.docs.amd.com/pro
 ### Added
 * `PipelineSerializer` class to implement pipeline serialization functionality in rocAL.
 * Python function execution support in rocAL.
+* Serialization test to validate pipeline serialization functionality.
 
 ### Changes
 * Adds new public APIs rocalSerialize() and rocalGetSerializedString() for serializing pipelines.
-* Add support to store the pipeline into protobuf format.
+* Add support to store the pipeline and introduce template-based serialization functions for different parameter types to convert to protobuf format.
 
 ### Removed